Component name improvement
