+++ src/views/movements.js
@@ -3,23 +3,58 @@ import { jsx } from "@emotion/core";
 import React from "react";
 import Footer from "../components/footer";
 import Header from "../components/header";
-import MovementForm from "../components/movement-form";
+import DayList from "../components/day-list";
+import { FaPlusCircle } from "react-icons/fa";
+import { navigate } from "@reach/router";
 
-function Movements() {
+function Movements({ isExpense }) {
+  function handleClick() {
+    navigate("/new");
+  }
+
   return (
-    <>
-      <Header title={"Register"} />
-      <MovementForm />
+    <section>
+      <Header title="Movements" />
+      <section
+        css={{
+          marginBottom: "3em",
+          overflow: "auto",
+          position: "relative"
+        }}
+      >
+        <DayList isExpense={isExpense} />
+      </section>
       <section
         css={{
           position: "fixed",
           bottom: "0px",
-          width: "100%"
+          right: "0px",
+          left: "0px"
         }}
       >
         <Footer />
       </section>
-    </>
+      <button
+        onClick={handleClick}
+        css={{
+          background: "none",
+          border: "none",
+          fontSize: "60px",
+          color: "#cc525c",
+          display: "flex",
+          position: "fixed",
+          bottom: "4rem",
+          right: "0.5rem",
+          backgroundColor: "#FFF",
+          borderRadius: "50%",
+          height: "60px",
+          width: "60px",
+          padding: "0"
+        }}
+      >
+        <FaPlusCircle />
+      </button>
+    </section>
   );
 }
